+++ src/inject-decorator.ts
@@ -1,0 +1,6 @@
+/**
+ * Inject decorator.
+ * 
+ * @type {InjectDecorator}
+ */
+export type InjectDecorator = (target: any, propertyName: string | symbol | undefined, injectIndex: number) => void;