+++ src/inject.ts
@@ -1,6 +1,7 @@
 import { isFunction, isNumber, isObject, isString, isUndefined } from 'lodash';
 import { isCtorFunction } from './functions/is-ctor-function';
 import { nameOf } from './functions/name-of';
+import { InjectDecorator } from './inject-decorator';
 import { InjectOptions } from './inject-options';
 import { TypeFn } from './type-fn';
 import { TypeManager } from './type-manager';
@@ -10,9 +11,9 @@ import { TypeManager } from './type-manager';
  * 
  * @param {TypeFn<TType>|InjectOptions<TType>|string} x Type function, inject options or parameter key from type context.
  * 
- * @returns {ParameterDecorator} Parameter decorator.
+ * @returns {InjectDecorator} Inject decorator.
  */
-export function Inject<TType>(x: TypeFn<TType> | InjectOptions<TType> | string): ParameterDecorator
+export function Inject<TType>(x: TypeFn<TType> | InjectOptions<TType> | string): InjectDecorator
 {
     const injectOptions = (isObject(x) ? x : {}) as InjectOptions<TType>;
 
@@ -25,8 +26,8 @@ export function Inject<TType>(x: TypeFn<TType> | InjectOptions<TType> | string):
     {
         injectOptions.typeFn = x as TypeFn<TType>;
     }
-
-    return function (target: any, propertyName: string | symbol, injectIndex: number): void
+    
+    return function (target: any, propertyName: string | symbol | undefined, injectIndex: number): void
     {
         if (!isCtorFunction(target))
         {