+++ src/property-decorator.ts
@@ -1,0 +1,6 @@
+/**
+ * Property decorator.
+ * 
+ * @type {PropertyDecorator}
+ */
+export type PropertyDecorator = (target: any, context: any) => any;