+++ src/type-decorator.ts
@@ -1,0 +1,6 @@
+/**
+ * Type decorator.
+ * 
+ * @type {TypeDecorator}
+ */
+export type TypeDecorator = (target: any, context?: any) => any;