+++ src/type.ts
@@ -1,3 +1,4 @@
+import { TypeDecorator } from './type-decorator';
 import { TypeFn } from './type-fn';
 import { TypeManager } from './type-manager';
 import { TypeOptions } from './type-options';
@@ -7,15 +8,18 @@ import { TypeOptions } from './type-options';
  * 
  * @param {TypeOptions<TType>} typeOptions Type options.
  *
- * @returns {ClassDecorator} Class decorator.
+ * @returns {TypeDecorator} Type decorator.
  */
-export function Type<TType>(typeOptions?: TypeOptions<TType>): ClassDecorator
+export function Type<TType>(typeOptions?: TypeOptions<TType>): TypeDecorator
 {
+    TypeManager.typeScope.open();
+
     return function (target: any): any
     {
         const typeFn = target as TypeFn<TType>;
+        const typeMetadata = TypeManager.configureTypeMetadata(typeFn, typeOptions).reflectInjectMetadata();
         
-        TypeManager.configureTypeMetadata(typeFn, typeOptions).reflectInjectMetadata();
+        TypeManager.typeScope.close(typeMetadata);
 
         return target;
     };