+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.2.1] - 2024-02-18
+
+### Fixed
+
+- TypeScript 5 decorators do not work.
+- Circular loop when using multiple inheritance through the decorator type options.
+
 ## [7.2.0] - 2023-10-18
 
 ### Added